flask-api/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.7
SQLAlchemy==2.0.19
paho-mqtt==1.6.1
python-dotenv==1.0.0
gunicorn==21.2.0
cryptography==41.0.4